refactor(chat): extract slash runtime parsing into shared module
